finished treating errors in all routes

+++ src/services/task.service.js
@@ -1,21 +1,46 @@
+const {
+  ApplicationError,
+  CommonHTTPExceptions,
+} = require("@pedromiotti/exerror");
+
+const { CustomExceptions } = require("../utils/exceptions/customExceptions");
+
+
 const TaskRepository = require("../repository/task.repository");
 
-async function createTask(description, status_id, user_id){
+async function createTask(description, status_id, user_id) {
+  if (!description || !status_id || !user_id)
+  throw new ApplicationError(CustomExceptions.MISSING_INFORMATION);
+
+  try {
     await TaskRepository.createTask(description, status_id, user_id);
+  } catch (e) {
+    throw new ApplicationError(e);
+  }
 }
 
-async function deleteTask(task_id){
+async function deleteTask(task_id) {
+  if (!task_id) throw new ApplicationError(CommonHTTPExceptions.BAD_REQUEST);
+
+  try {
     await TaskRepository.deleteTask(task_id);
+  } catch (e) {
+    throw new ApplicationError(e);
+  }
 }
 
-async function updateTask(task_id, description, status_id){
-    await TaskRepository.updateTask(task_id, description, status_id);
+async function updateTask(task_id, description, status_id, user_id) {
+  if (!task_id) throw new ApplicationError(CommonHTTPExceptions.BAD_REQUEST);
+
+  try {
+    await TaskRepository.updateTask(task_id, description, status_id, user_id);
+  } catch (e) {
+    throw new ApplicationError(e);
+  }
 }
-
-
 
 module.exports = {
-    createTask: createTask,
-    deleteTask:deleteTask,
-    updateTask:updateTask
-}
+  createTask: createTask,
+  deleteTask: deleteTask,
+  updateTask: updateTask,
+};
